+++ lib/domain/entities/address_entity.dart
@@ -1,5 +1,3 @@
-import 'dart:ffi';
-
 class AddressEntity {
   final String name;
   final String street;
@@ -8,8 +6,8 @@ class AddressEntity {
   final String state;
   final String country;
   final int zipcode;
-  final Long? latitude;
-  final Long? longitude;
+  final int? latitude;
+  final int? longitude;
 
   AddressEntity({
     required this.name,
